Create README.md
Made for learning the basics of React Native.
The user selects a number between 1 and 99, hit confirm and then start the game. The system will try to guess the number user chose. The user should press Greater or Lower button depending on the number the system guessed. At the end will show the number of rounds and the guessed number.